refactor: repository 패키지 위치를 도메인 패키지로 이동
